requirements.txt
pwinput
pyvmomi
python-dotenv
